requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-mptt~=0.13.3
asgiref==3.4.1
Babel==2.9.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-bootstrap4==3.0.1
django-js-asset==1.2.2
django-money==2.0.3
djangorestframework==3.12.4
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.2
psycopg2==2.9.1
py-moneyed==1.2
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.6
